Forward+ almost close to fucking final.
MISC:
-Reworked a bit Window->BeginFrame(), to return if failed to acquire an image.
-CommandBuffer can insert buffer memory barrier, added access flags.
-Timeline semaphores WIP.
VK:
Thanks to BEST-PRACTICES from validation layers, fixed SwapchainClearColor to be transfer op instead of graphics op.
Added vk_layer_settings.txt to make debugPrintfEXT working.
Fixed sampler minLod, maxLod, mipLoadBias, etc..
RENDERER:
-Switched to Reversed-Z
-Meshes fixed texture duplication while creating submeshes.
-Fixed FramebufferPerFrame, no need for them when using FIFs.
-ShaderLibrary added safe mutex when adding shaders to map since std::map is not thread safe?

TODO: Fix far/near plane check in light culling.
Kakoi zhe eto pizdec so vsemi etimi NDC, coordinate spaces kak zhe ya rot ebal bozhe naxyu.

+++ Pathfinder/Source/Platform/Vulkan/VulkanImage.cpp
@@ -249,8 +249,7 @@ void VulkanImage::Destroy()
 
     vkDestroyImageView(VulkanContext::Get().GetDevice()->GetLogicalDevice(), m_View, nullptr);
 
-    if (m_DescriptorInfo.sampler)
-        SamplerStorage::DestroySampler(SamplerSpecification{m_Specification.Filter, m_Specification.Wrap});
+    if (m_DescriptorInfo.sampler) SamplerStorage::DestroySampler(SamplerSpecification{m_Specification.Filter, m_Specification.Wrap});
 
     m_DescriptorInfo = {};
 
@@ -313,10 +312,13 @@ void* VulkanSamplerStorage::CreateSamplerImpl(const SamplerSpecification& sample
     samplerCI.addressModeV            = samplerCI.addressModeU;
     samplerCI.addressModeW            = samplerCI.addressModeU;
 
-    samplerCI.magFilter = VulkanUtility::PathfinderSamplerFilterToVulkan(samplerSpec.Filter);
-    samplerCI.minFilter = VulkanUtility::PathfinderSamplerFilterToVulkan(samplerSpec.Filter);
+    samplerCI.minLod     = samplerSpec.MinLod;
+    samplerCI.maxLod     = IsNearlyEqual(samplerSpec.MaxLod, .0f) ? VK_LOD_CLAMP_NONE : samplerSpec.MaxLod;
+    samplerCI.mipLodBias = samplerSpec.MipLodBias;
 
-    samplerCI.borderColor = VK_BORDER_COLOR_INT_OPAQUE_BLACK;  // TODO: configurable??
+    samplerCI.minFilter = samplerCI.magFilter = VulkanUtility::PathfinderSamplerFilterToVulkan(samplerSpec.Filter);
+
+    samplerCI.borderColor = VK_BORDER_COLOR_FLOAT_TRANSPARENT_BLACK;  // VK_BORDER_COLOR_INT_OPAQUE_BLACK;  // TODO: configurable??
     samplerCI.mipmapMode  = samplerCI.magFilter == VK_FILTER_NEAREST ? VK_SAMPLER_MIPMAP_MODE_NEAREST : VK_SAMPLER_MIPMAP_MODE_LINEAR;
 
     VkSampler sampler = VK_NULL_HANDLE;
